Fixes the capital letter issue
This was silly. The server kept crashing at certain moments ... half the
time. As it turns out, I forgot to convert the room code (and username)
to capital letters on the client side. As a result, the server couldn't
find the room it was looking for, because it was looking with the wrong
room code. This is fixed; everything should work smoothly.