Fixes not-progressing issue
I tried to get the length of an object (because it was, in older code,
an array). That doesn't work. (You need to get the keys of the object,
then get their length.)

Because of that, when all players had already submitted their
drawing/suggestion/guess, the game didn't automatically progress. Should
be fixed.