Finishes the game
Progress #23 Fixes #24
The game is now finished. It has all the functionality it needs and
everything (as far as I've tested) works as it should. The only thing it
needs is the finishing touch (prettier layout, animations, more
spectacular reveal, support for even the weirdest screen sizes, etc.).
Because this project was intended to be a stepping stone to an actual,
unique game, I don't know if I will ever execute that finishing touch.